divide coins
